moved CodeGen to separate file
